Moving ES specifics to test infrastructure subclasses.
