+++ graylog-storage-elasticsearch6/src/test/java/org/graylog/storage/elasticsearch6/CountsES6IT.java
@@ -6,6 +6,8 @@ import org.graylog2.indexer.counts.CountsAdapter;
 import org.graylog2.indexer.counts.CountsIT;
 import org.junit.Rule;
 
+import static org.graylog.storage.elasticsearch6.testing.TestUtils.jestClient;
+
 public class CountsES6IT extends CountsIT {
     @Rule
     public final ElasticsearchInstance elasticsearch = ElasticsearchInstanceES6.create();
@@ -17,6 +19,6 @@ public class CountsES6IT extends CountsIT {
 
     @Override
     protected CountsAdapter countsAdapter() {
-        return new CountsAdapterES6(jestClient());
+        return new CountsAdapterES6(jestClient(elasticsearch));
     }
 }
