+++ graylog-storage-elasticsearch6/src/test/java/org/graylog/storage/elasticsearch6/IndexFieldTypePollerES6IT.java
@@ -9,6 +9,8 @@ import org.graylog2.indexer.indices.IndicesAdapter;
 import org.graylog2.shared.bindings.providers.ObjectMapperProvider;
 import org.junit.Rule;
 
+import static org.graylog.storage.elasticsearch6.testing.TestUtils.jestClient;
+
 public class IndexFieldTypePollerES6IT extends IndexFieldTypePollerIT {
     @Rule
     public final ElasticsearchInstance elasticsearch = ElasticsearchInstanceES6.create();
@@ -21,11 +23,11 @@ public class IndexFieldTypePollerES6IT extends IndexFieldTypePollerIT {
     @Override
     protected IndicesAdapter createIndicesAdapter() {
         final ObjectMapper objectMapper = new ObjectMapperProvider().get();
-        return new IndicesAdapterES6(jestClient(), objectMapper, new IndexingHelper());
+        return new IndicesAdapterES6(jestClient(elasticsearch), objectMapper, new IndexingHelper());
     }
 
     @Override
     protected IndexFieldTypePollerAdapter createIndexFieldTypePollerAdapter() {
-        return new IndexFieldTypePollerAdapterES6(jestClient());
+        return new IndexFieldTypePollerAdapterES6(jestClient(elasticsearch));
     }
 }
