+++ graylog-storage-elasticsearch6/src/test/java/org/graylog/storage/elasticsearch6/testing/TestUtils.java
@@ -1,0 +1,14 @@
+package org.graylog.storage.elasticsearch6.testing;
+
+import io.searchbox.client.JestClient;
+import org.graylog.testing.elasticsearch.ElasticsearchInstance;
+
+public class TestUtils {
+    public static JestClient jestClient(ElasticsearchInstance elasticsearchInstance) {
+        if (elasticsearchInstance instanceof ElasticsearchInstanceES6) {
+            return ((ElasticsearchInstanceES6) elasticsearchInstance).jestClient();
+        }
+
+        throw new RuntimeException("Unable to return Jest client, Elasticsearch instance is of wrong type: " + elasticsearchInstance);
+    }
+}
