+++ graylog-storage-elasticsearch6/src/test/resources/elasticsearch.properties
